Updated strings to have 16 bit length

+++ libraries/WS12/Strings.h
@@ -7,9 +7,9 @@ class PStr {
   public:
     PStr(const char * buffer);
 
-    char operator[](uint8_t) const;
-    char operator[](uint8_t);
-    char charAt(uint8_t) const;
+    char operator[](uint16_t) const;
+    char operator[](uint16_t);
+    char charAt(uint16_t) const;
 
     bool equals(char * buffer, uint8_t length) const;
     bool begins(char * buffer, uint8_t length) const;
